+++ internal/usecase/delete_partner_usecase.go
@@ -1,6 +1,8 @@
 package usecase
 
 import (
+	"fmt"
+
 	"github.com/venture-technology/venture/internal/infra/contracts"
 	"github.com/venture-technology/venture/internal/infra/persistence"
 )
@@ -21,5 +23,15 @@ func NewDeletePartnerUseCase(
 }
 
 func (dpuc *DeletePartnerUseCase) DeletePartner(id string) error {
+
+	contracts, err := dpuc.repositories.ContractRepository.PartnerHasEnableContract(id)
+	if err != nil {
+		return err
+	}
+
+	if len(contracts) > 0 {
+		return fmt.Errorf("impossible delete partner because it has enable contract")
+	}
+
 	return dpuc.repositories.PartnerRepository.Delete(id)
 }
