+++ internal/usecase/delete_school_usecase.go
@@ -1,6 +1,8 @@
 package usecase
 
 import (
+	"fmt"
+
 	"github.com/venture-technology/venture/internal/infra/contracts"
 	"github.com/venture-technology/venture/internal/infra/persistence"
 )
@@ -21,5 +23,14 @@ func NewDeleteSchoolUseCase(
 }
 
 func (dsuc *DeleteSchoolUseCase) DeleteSchool(cnpj string) error {
+	SchoolHasContract, err := dsuc.repositories.ContractRepository.SchoolHasEnableContract(cnpj)
+	if err != nil {
+		return err
+	}
+
+	if SchoolHasContract {
+		return fmt.Errorf("impossivel deletar escola possuindo contrato ativo")
+	}
+
 	return dsuc.repositories.SchoolRepository.Delete(cnpj)
 }
